Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/stub-4.4-request-objects-expected.txt

CONSOLE MESSAGE: line 59: TypeError: null is not an object (evaluating 'document.getElementById("untested_idls").textContent')

FAIL Service Workers: Request Objects TypeError: null is not an object (evaluating 'document.getElementById("untested_idls").textContent')
